Profile frontend done

+++ TickTicket-Frontend/src/pages/Profile.js
@@ -1,15 +1,19 @@
 import React from "react";
-import {Button, Container, Stack, TextField} from "@mui/material";
+import {Alert, Button, Container, Stack, TextField} from "@mui/material";
 import axios from "axios";
 import { useEffect, useState } from "react";
 import ProfileCard from "../components/ProfileCard";
-// import EventCard from "../components/EventCard";
 
 const Home = () => {
     const [user, setUser] = useState();
     const [loading, setLoading] = useState(false);
     const [userId, setUserId] = useState(localStorage.getItem("userId"));
     const [password, setPassword] = useState("");
+    const [success, setSuccess] = useState(false);
+    const [successMessage, setSuccessMessage] = useState("");
+    const [error, setError] = useState(false);
+    const [errorMessage, setErrorMessage] = useState("");
+    const [visible, setVisible] = useState(false);
 
     const handlePasswordChange = (event) => {
         setPassword(event.target.value);
@@ -17,48 +21,45 @@ const Home = () => {
 
   useEffect(() => {
       console.log("user id from local storage is null? " + localStorage.getItem("userId") === "");
-      // loadData();
       setLoading(true);
-      axios
-          .get("users/" + localStorage.getItem("userId"))
-          .then((res) => {
-              setUser(res.data);
-              setUserId(localStorage.getItem("userId"));
-              console.log(res.data);
-              console.log(user);
-              console.log("user is null? " + user == null);
-              console.log("userID: " + user.id);
-              console.log(userId);
-          });
+      loadData();
       setLoading(false);
       // eslint-disable-next-line
   }, []);
 
-    // const loadData = async () => {
-    //     setLoading(true);
-    //     await axios
-    //         .get("users/" + localStorage.getItem("userId"))
-    //         .then((res) => {
-    //             setUser(res.data);
-    //             setUserId(localStorage.getItem("userId"));
-    //             // console.log(res.data);
-    //             console.log(user);
-    //             console.log("user is null? " + user == null);
-    //             console.log("userID: " + user.id);
-    //             console.log(userId);
-    //         });
-    //     setLoading(false);
-    // }
+    const loadData = () => {
+        setLoading(true);
+        axios
+            .get("users/" + localStorage.getItem("userId"))
+            .then((res) => {
+                setUser(res.data);
+                setUserId(localStorage.getItem("userId"));
+                console.log(userId);
+            });
+        setLoading(false);
+    }
 
     const handleDelete = React.useCallback(() => {
-        console.log("Delete button was pressed");
-        axios.delete("users/" + userId, {
-            params: {
-                password: password
-             }
-    })
 
-    }, [userId, password])
+        setVisible(true);
+        console.log("Delete button was pressed");
+        if(visible) {
+            axios.delete("users/" + userId, {
+                params: {
+                    password: password
+                }
+            }).then(res => {
+                setSuccess(true);
+                setSuccessMessage("Account was deleted successfully");
+                localStorage.removeItem("userId");
+                window.location.href = "/";
+            }).catch(function (error) {
+                setError(true);
+                setErrorMessage("Account could not be deleted.");
+            });
+        }
+        console.log(errorMessage)
+    }, [userId, password, visible, errorMessage])
 
   return (
       <Container
@@ -69,21 +70,21 @@ const Home = () => {
           <Stack spacing={2}>
               <h1>Profile</h1>
               {!loading && user != null && (
-                  <ProfileCard key={user.id} event={user}/>)
+                  <ProfileCard key={user.id} user={user}/>)
               }
           </Stack>
           <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', marginTop: 20}}>
-              <TextField  type="password" label="password" variant="standard" onChange={handlePasswordChange}/>
+              { visible && <TextField type="password" label="password" variant="standard" onChange={handlePasswordChange}/>}
           </div>
           <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', marginTop: 20}}>
-              <Button
-                  onClick={handleDelete}
-                  color="primary"
-              >
+              <Button onClick={handleDelete} color="primary">
                   DELETE ACCOUNT
               </Button>
           </div>
-
+          <Stack sx={{ width: '100%' }} spacing={2}>
+              {error && <Alert severity="error">{errorMessage}</Alert>}
+              {success && <Alert severity="success">{successMessage}</Alert>}
+          </Stack>
       </Container>
   );
 };
